auth fix

+++ Group5/routes/web.php
@@ -22,16 +22,7 @@ Route::get('/', function () {
     return view('index');
 });
 
-// Route::middleware([
-//     'auth:sanctum',
-//     config('jetstream.auth_session'),
-//     'verified',
-// ])->group(function () {
-//     Route::get('/dashboard', function () {
-//         $users = User::all();
-//         return view('dashboard', compact('users'));
-//     })->name('dashboard');
-// });
+
 Route::get('/contact', function () {
     $users = User::all();
     return view('contact');
@@ -52,37 +43,36 @@ Route::get('/register-view', function(){
     return view('auth.register');
 });
 
-//Admin CRUD
+//users CRUD
+
 Route::get('/dashboard', function () {
     $users = User::all();
     return view('dashboard', compact('users'));
 })->name('dashboard')->middleware('role:admin');
-Route::get('/users/{id}/edit',[UserController::class ,'edit'])->name('users.edit')->middleware('role:admin');
-Route::put('/users/{id}', [UserController::class ,'update'])->name('users.update')->middleware('role:admin');
-Route::get('/users/{id}', [UserController::class ,'show'])->name('users.show')->middleware('role:admin');
-Route::delete('/users/{id}', [UserController::class ,'destroy'])->name('users.destroy')->middleware('role:admin');
+Route::get('/users/{id}/edit',[UserController::class ,'edit'])->name('users.edit')->middleware('role:admin')->middleware('role:admin');
+Route::put('/users/{id}', [UserController::class ,'update'])->name('users.update')->middleware('role:admin')->middleware('role:admin');
+Route::get('/users/{id}', [UserController::class ,'show'])->name('users.show')->middleware('role:admin')->middleware('role:admin');
+Route::delete('/users/{id}', [UserController::class ,'destroy'])->name('users.destroy')->middleware('role:admin')->middleware('role:admin');
 
 
 //Employer
 Route::get('/jobs', [JobController::class, 'index'])->name('job.index')->middleware('role:employer');
 Route::get('/jobs/create', [JobController::class, 'create'])->name('job.create')->middleware('role:employer');
 Route::post('/store', [JobController::class, 'store'])->name('job.store')->middleware('role:employer');
-Route::get('/job/{job}/edit', [JobController::class, 'edit'])->name('job.edit');
-
-Route::put('/job/{job}/update', [JobController::class, 'update'])->name('job.update');
-Route::delete('/job/{job}/delete', [JobController::class, 'delete'])->name('job.delete');
-Route::get('/job/{job}/applicants', [JobController::class, 'show_applicants'])->name('job.applicants');
+Route::get('/job/{job}/edit', [JobController::class, 'edit'])->name('job.edit')->middleware('role:employer');
+Route::put('/job/{job}/update', [JobController::class, 'update'])->name('job.update')->middleware('role:employer');
+Route::delete('/job/{job}/delete', [JobController::class, 'delete'])->name('job.delete')->middleware('role:employer');
+Route::get('/job/{job}/applicants', [JobController::class, 'show_applicants'])->name('job.applicants')->middleware('role:employer');
 
 //applicant
-Route::get('/applicant/jobs', [JobController::class, 'showAllJobsToApplicant'])->name('applicant.index');
-Route::get('/applicant/{job}/apply', [ApplicationController::class, 'apply'])->name('applicant.apply');
-Route::post('/applicant/submit-application', [ApplicationController::class, 'submit_application'])->name('applicant.submit');
-Route::get('/applicant/my-applications', [ApplicationController::class, 'my_applications'])->name('applicant.my-applications');
-Route::delete('/applicant/{application}/cancel', [ApplicationController::class, 'cancel'])->name('applicant.cancel');
+Route::get('/applicant/jobs', [JobController::class, 'showAllJobsToApplicant'])->name('applicant.index')->middleware('role:applicant');
+Route::get('/applicant/{job}/apply', [ApplicationController::class, 'apply'])->name('applicant.apply')->middleware('role:applicant');
+Route::post('/applicant/submit-application', [ApplicationController::class, 'submit_application'])->name('applicant.submit')->middleware('role:applicant');
+Route::get('/applicant/my-applications', [ApplicationController::class, 'my_applications'])->name('applicant.my-applications')->middleware('role:applicant');
+Route::delete('/applicant/{application}/cancel', [ApplicationController::class, 'cancel'])->name('applicant.cancel')->middleware('role:applicant');
 
-
-Route::get('/all/category', [CategoryController::class,'index'])->name('AllCat')->middleware('role:admin');
-Route::post('/all/category', [CategoryController::class, 'store'])->name('categories.store')->middleware('role:admin');
-Route::get('/category/edit/{id}', [CategoryController::class, 'Edit'])->middleware('role:admin');
-Route::post('/category/update/{id}', [CategoryController::class, 'Update'])->name('update.category')->middleware('role:admin');
-Route::get('/category/delete/{id}', [CategoryController::class, 'Delete'])->middleware('role:admin');
+Route::get('/all/category', [CategoryController::class,'index'])->name('AllCat')->middleware('role:applicant');
+Route::post('/all/category', [CategoryController::class, 'store'])->name('categories.store')->middleware('role:applicant');
+Route::get('/category/edit/{id}', [CategoryController::class, 'Edit'])->middleware('role:applicant');
+Route::post('/category/update/{id}', [CategoryController::class, 'Update'])->name('update.category')->middleware('role:applicant');
+Route::get('/category/delete/{id}', [CategoryController::class, 'Delete'])->middleware('role:applicant');
